fix: renovate parsing jinja2 yaml incorrectly
Recently, a PR was merged in the Renovate repository: https://github.com/renovatebot/renovate/pull/31206
This was supposed to implement Jinja2 support for YAML files, but actually backfired in my case since it seems that the regex used is incorrect, and would think the template never ends and thus take the whole file with it.
It seems that wrapping the templates in quotation marks fixes the issue.